==> 1_NumberOfMentions.ipynb <==
import csv
import pandas as pd
import datetime as dt
import json
import time
from urllib.request import urlopen

This script counts the number of mentions on WallStreetBets for every single stock across the 3 major US stock exchanges - this is important in determining which stocks to use and which to ommit, as the threshold is a minimum of 100 mentions for the stock to be included in the sample

with open(r'C:\Users\Nish\OneDrive - University of Bristol\Dissertation\Volume Forecasting\Data\Tickers\Alltickers.csv') as csv_file:
    reader = csv.reader(csv_file, delimiter=',')
    ticker = list(reader)
ticklist = []
for sublist in ticker:
    for item in sublist:
        ticklist.append(item)
url_list=[]
for i in ticklist:
    url_list.append('https://api.pushshift.io/reddit/search/submission/?q={tick}&subreddit=wallstreetbets&metadata=true&size=0&after=1610928000&before=1613088000'.format(tick=i))

import sqlite3
conn = sqlite3.connect('tickers.db')
conn.execute('''CREATE TABLE DB(URL, TICKER, STATUS default "Not started");''')

for i, j in zip(url_list,ticklist):
    conn.execute("INSERT INTO DB(URL, TICKER) VALUES(?,?)",(i,j))

#for row in conn.execute("SELECT URL, TICKER, STATUS from DB"):
    #print('URL={i}'.format(i=row[0]))
    #print('TICKER={i}'.format(i=row[1]))
    #print('STATUS={i}'.format(i=row[2]))

X = []

for row in conn.execute("SELECT URL, TICKER, STATUS from DB"): 
    if row[2] == 'Not started':
        f = urlopen("https://api.pushshift.io/reddit/search/submission/?q={tick}&subreddit=wallstreetbets&metadata=true&size=0&after=1610928000&before=1613088000".format(tick=row[1]))
        j = json.load(f)
        subs = j['metadata']['total_results']
        X.append(subs)
        print('{tick} has been scraped!'.format(tick=row[1]))
        conn.execute("UPDATE DB set STATUS = 'Scraped' WHERE TICKER = ?",(row[1],))
        time.sleep(1)
    else:
        continue 

df1 = pd.DataFrame(X,columns=['Mentions'])
df2 = pd.DataFrame(ticklist, columns=['Stock'])
df3 = pd.concat([df2.reset_index(drop=True),df1.reset_index(drop=True)], axis=1)

df3.to_csv(r"C:\Users\Nish\OneDrive - University of Bristol\Dissertation\Volume Forecasting\Data\Tickers\AllStocks.csv")
