==> 2_StockDataScrape.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import requests
import csv
from sys import platform
from datetime import date
import time
key1 = "9DIS4V39ATIST4CK"

This script uses the AlphaVantage API to scrape the financial data for each stock in the sample

#importing the tickers of interest
if platform == 'darwin':    
    tick = pd.read_csv(r'/Users/nishaldave/OneDrive - University of Bristol/Dissertation/Volume Forecasting/Data/tickers',header=None)
    ticklist = []
    for i in tick[0]:
      ticklist.append(i)
elif platform == 'win32':
    tick = pd.read_csv(r'C:\Users\Nish\OneDrive - University of Bristol\Dissertation\Volume Forecasting\Data\Tickers\AVScrape\avtickers.csv', header=None)
    ticklist = []
    for i in tick[0]:
      ticklist.append(i)

#Scraping financial data from the Alpha-Vantage API
from alpha_vantage.timeseries import TimeSeries
ts = TimeSeries(key=key1, output_format='pandas')
stonkspanel = []
for count,i in enumerate(ticklist):
  data, meta_data = ts.get_daily_adjusted(symbol=i, outputsize='full')
  data['asset']  = i
  stonkspanel.append(data)
  print('{x} has been scraped'.format(x=i))
  if (count+1)%25==0:
    time.sleep(65)
  else:
      continue

#Formatting stock data into datetime format
stocks = pd.concat(stonkspanel)
stocks = stocks.reset_index()
stocks['week'] = stocks['date'].dt.week
stocks['date'] = pd.to_datetime(stocks['date'],dayfirst=True).dt.date

#Time period used is from 29/06/2020 to 25/06/2021 - 1 year or 52 whole weeks (This is important when I use weekly data)
stocks = stocks[stocks['date']>date(2020, 6, 25)]
stocks = stocks[stocks['date']<date(2021, 6, 26)]
#Renaming the variables and dropping irrelevant variables
stocks = stocks.drop(columns=['5. adjusted close','7. dividend amount', '8. split coefficient'])
stocks = stocks.rename(columns={'1. open' : 'open', '2. high' : 'high', '3. low' : 'low', '4. close' : 'close', '6. volume' : 'volume'})

#Coding the weeks as ascending integers

week_map_a = {(i+26):i for i in range(1,28)}
week_map_b = {i:i+27 for i in range(1,27)}
week_map  = {a:b for a,b in list(week_map_a.items())+list(week_map_b.items())}
stocks['week'] = stocks['week'].apply(lambda x: week_map[x])
#Sort by descending date
stocks = stocks.iloc[::-1]
#Save CSV to local
if platform == 'darwin':
    pd.DataFrame.to_csv(stocks,path_or_buf='/content/drive/MyDrive/Dissertation/Volume Forecasting/Data/stocks2.csv')
elif platform == 'win32':
    pd.DataFrame.to_csv(stocks,path_or_buf=r'C:\Users\Nish\OneDrive - University of Bristol\Dissertation\Volume Forecasting\Data\Stock Scraper\stocks_updated.csv')

len(stocks['asset'].unique())
