==> 3_WSBscraper.ipynb <==
from psaw import PushshiftAPI
import pandas as pd
import numpy as np
import datetime as dt
from sys import platform
api = PushshiftAPI()
pd.options.mode.chained_assignment = None  # default='warn'

This script scrapes WallStreetBets using the Pushshift API and collects data related to each submission on the forum regarding each stock

#Easy switching between Windows and Mac directories
if platform == 'darwin':
    ticker = pd.read_csv(r'/Users/nishaldave/OneDrive - University of Bristol/Dissertation/Volume Forecasting/Data/WSB Scraper/tickers2.csv',header=None)
    ticker[1] = '$'+ticker[0]
    ticker.to_csv(r'/Users/nishaldave/OneDrive - University of Bristol/Dissertation/Volume Forecasting/Data/WSB Scraper/tickers2_new.csv', index=False)
    import csv
    with open(r'/Users/nishaldave/OneDrive - University of Bristol/Dissertation/Volume Forecasting/Data/WSB Scraper/tickers2_new.csv') as csv_file:
        reader = csv.reader(csv_file, delimiter=',')
        tickers = list(reader)
    del tickers[0]
    
elif platform =='win32':
    ticker = pd.read_csv(r'C:\Users\Nish\OneDrive - University of Bristol\Dissertation\Volume Forecasting\Data\Tickers\TickersToScrape.csv',header=None)
    #ticker[1] = '$'+ticker[0]
    ticker.to_csv(r'C:\Users\Nish\OneDrive - University of Bristol\Dissertation\Volume Forecasting\Data\Tickers\TickersToScrape2.csv', index=False)
    import csv
    with open(r'C:\Users\Nish\OneDrive - University of Bristol\Dissertation\Volume Forecasting\Data\tickers\TickersToScrape2.csv') as csv_file:
        reader = csv.reader(csv_file, delimiter=',')
        tickers = list(reader)
    del tickers[0]

ticklist = []
for sublist in tickers:
    for item in sublist:
        ticklist.append(item)

import sqlite3
conn = sqlite3.connect('WSBScraper.db')
conn.execute('''CREATE TABLE DB(TICKER, STATUS default "Not started");''')

for i in ticklist:
    conn.execute("INSERT INTO DB(TICKER) VALUES(?)",(i,))

for row in conn.execute("SELECT TICKER, STATUS from DB"):
    print('TICKER={i}'.format(i=row[0]))
    print('STATUS={i}'.format(i=row[1]))

dfs = []

#Scrapes all reddit submissions associated with each ticker 

for row in conn.execute("SELECT TICKER, STATUS from DB"):
        if row[1] == 'Not started':

            start_date = int(dt.datetime(2020, 6, 29, 0,0,0).timestamp())
            start_date2 = int(dt.datetime(2020, 6, 29, 23,59,59).timestamp())
            end_date = int(dt.datetime(2021, 6, 25, 23,59,59).timestamp())

            gen_init = api.search_submissions(subreddit='wallstreetbets',q=row[0],limit=10000,after=start_date,before=start_date2)
            df_init = pd.DataFrame([X.d_ for X in gen_init])
            df_final = df_init[0:0]

            while start_date < end_date:    
              gen_data = api.search_submissions(subreddit='wallstreetbets',q=row[0],limit=10000,after=start_date,before=start_date2)
              df_inter = pd.DataFrame([X.d_ for X in gen_data])
              df_final = pd.concat([df_final,df_inter],axis=0)
              #df_final['created_utc'] = pd.to_datetime(df_final['created_utc'],unit='s')
              df_final['asset'] = row[0]
              start_date  = start_date  + 86400
              start_date2 = start_date2 + 86400
            
            dfs.append(df_final)
            print('{tick} has been fully scraped!'.format(tick=row[0]))
            conn.execute("UPDATE DB set STATUS = 'Scraped' WHERE TICKER = ?",(row[0],))            
        else:
            continue

start_date = int(dt.datetime(2020, 6, 29, 0,0,0).timestamp())
start_date2 = int(dt.datetime(2020, 6, 29, 23,59,59).timestamp())
end_date = int(dt.datetime(2021, 6, 25, 23,59,59).timestamp())

gen_init = api.search_submissions(subreddit='wallstreetbets',q='$F',limit=1,after=start_date,before=start_date2)
df_init = pd.DataFrame([X.d_ for X in gen_init])
df_final = df_init[0:0]

while start_date < end_date:    
  gen_data = api.search_submissions(subreddit='wallstreetbets',q='$F',limit=10000,after=start_date,before=start_date2)
  df_inter = pd.DataFrame([X.d_ for X in gen_data])
  df_final = pd.concat([df_final,df_inter],axis=0)
  #df_final['created_utc'] = pd.to_datetime(df_final['created_utc'],unit='s')
  df_final['asset'] = '$F'
  start_date  = start_date  + 86400
  start_date2 = start_date2 + 86400

dfs.append(df_final)

#Necessary variables to retain
#created_utc, link_flair_css_class, num_comments, selftext, title, asset.

dfs[0] = dfs[0][['created_utc','link_flair_css_class','num_comments','selftext','title','asset']]
dfs[0]['created_utc'] = pd.to_datetime(dfs[0]['created_utc'],unit='s')
dfs[0] = dfs[0].sort_values(by='created_utc',ascending=True)
dfs[0] = dfs[0][dfs[0]['selftext'] != '[removed]']
dfs[0] = dfs[0][dfs[0]['selftext'] != '[deleted]']
dfs[0] = dfs[0][dfs[0]['selftext'].notna()]
dfs[0] = dfs[0][dfs[0]['selftext']!=''] #drop selftexts where there is no valid description.

delete all entries from ford - then scrape $F ticker,
delete all entries from JP and delete JP stock info if there is any

#Necessary variables to retain
#created_utc, link_flair_css_class, num_comments, selftext, title, asset.
for i in range(len(dfs)):
    dfs[i] = dfs[i][['created_utc','link_flair_css_class','num_comments','selftext','title','asset']]
    dfs[i]['created_utc'] = pd.to_datetime(dfs[i]['created_utc'],unit='s')
    dfs[i] = dfs[i].sort_values(by='created_utc',ascending=True)
    dfs[i] = dfs[i][dfs[i]['selftext'] != '[removed]']
    dfs[i] = dfs[i][dfs[i]['selftext'] != '[deleted]']
    dfs[i] = dfs[i][dfs[i]['selftext'].notna()]
    dfs[i] = dfs[i][dfs[i]['selftext']!=''] #drop selftexts where there is no valid description.

pd.concat(dfs).to_csv(r'C:\Users\Nish\OneDrive - University of Bristol\Dissertation\Volume Forecasting\Data\WSB Scraper\Submissions\assetsF.csv')

#pd.concat(dfs).to_csv('/Users/nishaldave/OneDrive - University of Bristol/Dissertation/Volume Forecasting/Data/WSB Scraper/Submissions/assets3_v1.csv')
