==> 4_WSBcleaning.ipynb <==
import pandas as pd
import datetime
import nltk
nltk.download('vader_lexicon')
from nltk.sentiment.vader import SentimentIntensityAnalyzer
analyzer = SentimentIntensityAnalyzer()

This script cleans and manages the Reddit data as well as computing the VADER sentiment values for each submission

df = pd.read_csv(r'C:\Users\Nish\OneDrive - University of Bristol\Dissertation\Volume Forecasting\Data\WSB Scraper\Submissions\assets3_v1.csv')
df2 = pd.read_csv(r'C:\Users\Nish\OneDrive - University of Bristol\Dissertation\Volume Forecasting\Data\WSB Scraper\Submissions\assetsF.csv')

df = pd.concat([df,df2])

df = df.reset_index()
#Dropping the asset S (copy of SPY, not sure why it is there)
#df = df[df['asset']!= 'S']

#Add the week of the year
df['created_utc'] =  pd.to_datetime(df['created_utc'])
df['week'] = df['created_utc'].dt.week
df['date'] = pd.to_datetime(df['created_utc'],dayfirst=True).dt.date

#Coding the weeks as ascending integers
week_map_a = {(i+26):i for i in range(1,28)}
week_map_b = {i:i+27 for i in range(1,27)}
week_map  = {a:b for a,b in list(week_map_a.items())+list(week_map_b.items())}
df['week'] = df['week'].apply(lambda x: week_map[x])

df['compound1'] = df['selftext'].apply(lambda x:analyzer.polarity_scores(x)['compound'])
#df['compound2'] = df['title'].apply(lambda x:analyzer.polarity_scores(x)['compound'])

df = df[df['asset'] != '$JP']

df['asset'] = df['asset'].map(lambda x: x.lstrip('$'))

pd.DataFrame.to_csv(df,path_or_buf=r'C:\Users\Nish\OneDrive - University of Bristol\Dissertation\Volume Forecasting\Data\WSB Scraper\Submissions\assets3_v2.csv')
